Converts code to typescript
