+++ extension/events.ts
@@ -1,3 +1,4 @@
+import { NodeCG, Replicant } from "nodecg-types/types/server"
 /*
 Event logger: Keeps a history of all events that happened
 during the Race Against Time.
@@ -5,9 +6,20 @@ during the Race Against Time.
 const TIMESTAMP_REGEX = /At$/;
 const copy = (obj) => JSON.parse(JSON.stringify(obj));
 
-module.exports = (nodecg) => {
+export interface Event {
+  id: string;
+  icon?: string;
+  title: string;
+  description?: string;
+  shown?: boolean;
+  occuredAt: Date;
+  // TODO: add 'type'
+  [x: string]: any; // Optionally supports other properties.
+}
+
+export default (nodecg: NodeCG) => {
   nodecg.Replicant("events-bar", { defaultValue: "" });
-  const events = nodecg.Replicant("events", { defaultValue: [] });
+  const events: Replicant<Array<Event>> = nodecg.Replicant("events", { defaultValue: [] });
 
   nodecg.listenFor("event", (e) => {
     const newEvent = copy(e);
@@ -17,7 +29,7 @@ module.exports = (nodecg) => {
     // while the name of the key may differ, all timestamps match the pattern
     // ____At (i.e. occuredAt, achievedAt, etc.).
     // Take the first match and use that as the occuredAt timestamp.
-    const timestamp = Object.keys(newEvent).find((key) =>
+    const timestamp: any = Object.keys(newEvent).find((key) =>
       TIMESTAMP_REGEX.test(key)
     );
     events.value.push({
