--- extension/start.js
@@ -1,3 +1,0 @@
-module.exports = (nodecg) => {
-  nodecg.Replicant("start", false);
-};
