+++ package.json
@@ -38,6 +38,7 @@
     "eslint-plugin-node": "^11.1.0",
     "eslint-plugin-promise": "^4.3.1",
     "eslint-plugin-react": "^7.22.0",
+    "nodecg-types": "^1.8.3",
     "parcel": "^2.0.0",
     "preact": "^10.0.0",
     "prettier": "^2.2.1",
