+++ src/dashboard/achievements.tsx
@@ -1,8 +1,9 @@
+/// <reference types="nodecg-types/types/browser"/>
 import { h, render, FunctionComponent, Fragment } from "preact";
 import { useCallback, useState } from "preact/hooks";
 import { useReplicant } from "use-nodecg";
-import type { Achievement } from "../../types/replicants";
-import type { Completeable } from "../../types/events";
+import type { Achievement } from "../../extension/achievements";
+import type { Completeable } from "../../extension/checkpoints";
 
 const copy = (obj: any) => JSON.parse(JSON.stringify(obj));
 
@@ -50,7 +51,7 @@ const AchievementsPanel: FunctionComponent<any> = () => {
           if (!achievement.achievedAt) {
             achievement.achievedAt = new Date();
           }
-          (window as any).nodecg.sendMessage("event", achievement);
+          nodecg.sendMessage("event", achievement);
         }
         return true;
       });
