--- types/events.d.ts
@@ -1,8 +1,0 @@
-// TODO: Figure out how to fix unused vars for typescript like this.
-export interface Completeable {
-  onComplete: (id: string, event: any) => void; // eslint-disable-line no-unused-vars
-}
-
-export interface Changeable {
-  onChange: (id: string, event: any) => void; // eslint-disable-line no-unused-vars
-}
